test: add infra for running WCT tests
Summary:
Currently, we have some TypeScript unit tests, but we don’t run them
because they can only be run in a browser. This commit adds
infrastructure for running these tests headlessly and automatically.

Test failures will unfortunately simply be printed as

    AssertionError: 1 failing test

because it is not obvious how to easily get additional information about
which tests have failed and why.

After running `bazel build`, running `bazel test` takes 17–18 seconds on
a 6-core W-2135 workstation, which is obviously obscene and untenable.
Running with `bazel run` takes “only” about 5s, which is still obscene
and untenable, but raises the question of why there is such a large
difference. We will need to investigate this performance if we want to
actually use this pattern to run our tests.

Test Plan:
Run `bazel test //tensorboard/components/vz_sorting/test` and note that
it passes. Then, change `sortingTests.ts` to make one of the expected
outputs incorrect; rerun the test command and note that it fails. Note
that at no time does a browser window appear.

wchargin-branch: test-wct-infra